Add Seedig of database

+++ db/migrate/20170725231556_create_votes.rb
@@ -6,9 +6,9 @@ class CreateVotes < ActiveRecord::Migration[5.1]
       t.string :validity
       t.string :choice
       t.string :conn
-      t.integer :msisdn
+      t.bigint :msisdn
       t.string :guid
-      t.integer :short_code
+      t.bigint :short_code
 
       t.timestamps
     end
